kbuild: add drivers/video/tegra/host/Kconfig
In the nvidia overlay, drivers/video/tegra/Kconfig includes
drivers/video/tegra/host/Kconfig. The file doesn't currently exist in the
nvidia overlay, despite the directory existing and containing a non-empty
Makefile and source code. The file does exist in other overlays, and
contains necessary content there. Add an empty file so that this include
statement succeeds even if the other overlay(s) is/are not present.

An alternative would be to modify the other overlay(s) so that a parent
directory's Kconfig includes this Kconfig file. However, that approach is
more applicable to entire directories that only exist in the other
overlay(s); since this host/ directory exists in the nvidia overlay, and
contains other meaninful content, I think it should contain a base version
of all infrastructure files.

Bug 1978395
Bug 2015191

Change-Id: I3a26c1a301a89a65b820a230ee21fd44c5271986
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1588587
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>