t18x: soc: align SOR clk DT names
Currently, the DT names that we're using for the SOR clks
are different on T21x vs. T186/T194, which forces us to
unnecessarily maintain divergent paths in the SOR driver
itself. In order to rectify this problem, this patch updates
the T186 SOC base DT file so that it follows this naming scheme:
- "sorX_ref" : SOR reference clk that's output by the
               SORX_CLK_SRC MUX
- "sor_safe" : SOR safe clk
- "sorX_pad_clkout" : clk signal output by the SORX pad brick
- "sorX" : clk signal output by the SORX_CLK_SEL0 MUX

Change-Id: I6da26e81440a3bb4d52ed3742263f8aa06fdb3c4
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1589692
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>