t19x: soc: align SOR clk DT names
Currently, the DT names that we're using for the SOR clks
are different on T21x vs. T186/T194, which forces us to
unnecessarily maintain divergent paths in the SOR driver
itself. In order to rectify this problem, this patch updates
the T194 SOC base DT file so that it follows this naming scheme:
- "sorX_ref" : SOR reference clk that's output by the
               SORX_CLK_SRC MUX
- "sor_safe" : SOR safe clk
- "sorX_pad_clkout" : clk signal output by the SORX pad brick
- "sorX" : clk signal output by the SORX_CLK_SEL0 MUX

In addition, this patch fixes the incorrect clock IDs that are
currently used for "sorX_ref" and "sorX", and also adds the
missing entries for PLLD3 and PLLD4.

TDS-2839

Change-Id: Iff6c209a6ad30ca9badfe2a7f58343258e86814b
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1589693
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>