dc: unify SOR clk names
The base DT files for each SOC are being updated to follow
this naming scheme for the SOR clks:
- "sorX_ref" : SOR reference clk that's output by the
               SORX_CLK_SRC MUX
- "sor_safe" : SOR safe clk
- "sorX_pad_clkout" : clk signal output by the SORX pad brick
- "sorX" : clk signal output by the SORX_CLK_SEL0 MUX

This patch changes the SOR clk field names to more closely align
with the DT names, and unifies the clk logic in the SOR init
path across T21x vs. T186/T194.

In addition, the naming and usage of the old sor->src_switch_clk
and sor->sor_clk fields were completely swapped on T186/T194 for
no good reason - this change fixes that as well.

TDS-2839

Change-Id: I9d47a5cfd15fe4d6e84d7521482975e83c0e784a
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1589688
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>