se-elp: Remove health check from RNG1 code
Remove NOISE_RDY health check from the driver
as this code does not know if it is the first piece
of SW to trigger RNG1 HW or not.
This should be ideally checked before any of the SW
actually uses it, which is MB1. This is being discussed
in Bug 200356385.

Change-Id: Iacf42b3b35c947165b591ea9653d07e6c6ceb536
Signed-off-by: Shravani Dingari <shravanid@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1590680
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mallikarjun Kasoju <mkasoju@nvidia.com>
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>