platform: tegra: pm: create 32-bit specific wk_test_bit()
test_bit() operates on ulong, which on a 64-bit arch is 64-bits.
test_bit() is however fed with a 3*32-bit array,
which means that when bits are set/cleared in the last element
test_bit() willl access memory outside of the defined 3*32-bit array.

Instead of using test_bit(), roll our own 32-bit specific function.

Bug 200342601

Change-Id: Iad3dc6e50fda5feea0dfa26eaa2e719814fb0823
Signed-off-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1595975
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>