drivers:csi:csi2 error reset fix
The csi port code had changed so that only one port
is allocated to a channel no matter how many physical
ports are used. The csi2 error handler was not updated
for this change and was trying to set 2 ports when only
one existed. Change updates the csi to use the port offset
values to index the second csi to clear the buffer after an
error occurs.

Bug 1992619

Change-Id: If29a15e67ad8cd507d4aa7a03113ba616610e3cf
Signed-off-by: Josh Kergan <jkergan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1596406
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jihoon Bang <jbang@nvidia.com>
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>