t18x: mce: Ignore T21x platforms
On T210 platforms, there are some errors:
  ERROR:   tegra_sip_handler: unhandled SMC (0xc2ffff0e)
  [  398.411593] tegra_roc_flush_cache_only: failed (ret=-1)
  ERROR:   tegra_sip_handler: unhandled SMC (0xc2ffff0f)
  [  398.505041] tegra_roc_clean_cache: failed (ret=-1)
  ERROR:   tegra_sip_handler: unhandled SMC (0xc2ffff0e)
  [  398.517037] tegra_roc_flush_cache_only: failed (ret=-1)
  ERROR:   tegra_sip_handler: unhandled SMC (0xc2ffff0f)
  [  398.602320] tegra_roc_clean_cache: failed (ret=-1)

This tegra_roc_flush_cache_only() function in the tegra-mce driver
is supposed to be run on T186/T194 only. However, K4.9 has unified
the defconfig for both T21x and T18x so the tegra-mce driver with
its functions is being compiled on T210 platforms as well.

So this patch adds an SoC check to ignore T21x platforms to fix it.

Bug 200363383

Change-Id: I0ff38aad22b183615e87e8d6be21511ff7187aa2
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1597453
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>