gpio: t186: Add support for ignoring GPIO int route
T19x GPIO controller supports 8 interrupt line to generate
interrupt to system. The pins of GPIO controller can be map
to any of interrupt line.

Add support to ignore the interrupt registration for some of
the interrupt line if the required information is provided from
DT. On this way, system will not register the interrupt for
the CPU for that interrupt line.

Bug 200351598

Change-Id: If5a9b4af1b84937e0c167c9b4a11d9937b44d36e
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1601623
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Pavan Kunapuli <pkunapuli@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>