platform: tegra: Reconstruct mce and cache drivers
The tegra-mce driver was designed for T18x only and being selected
based on CONFIG_ARCH_TEGRA_18x_SOC. But Kernel 4.9 uses a unified
defconfig file which means CONFIG_ARCH_TEGRA_18x_SOC is enabled on
platforms like T21x as well. The same thing happends to the system
with CONFIG_ARCH_TEGRA_19x_SOC as well.

This creates a problem that: with previous kernel version the mce
driver was not expected to be compiled while now it's compiled and
its functions are being called by other drivers without caring if
it's fine to execute them or not.

Althouth there are some code being added to the functions based on
the SoC ID so as to make them be NOP functions for older platforms
like T21x, this approach is not only tedious because every exported
function would need an ID check, but also doesn't fix all issues.

The cache functions defined in the tegra-mce driver are using the
same names of those default functions in the arch/arm64/mm/cache.S
file, which means, once the tegra-mce driver gets compiled, these
mce cache functions will override the default ones. On a platform
like T21x, it is supposed to call those default functions but now
they gets NOP functions instead.

So this patch reconstructs the whole mce and cache drivers:
1) Changes tegra-mce driver to be a core driver that runs on all
   platforms and defines corresponding function pointers for the
   running platform: Abstracts similar functions in this file and
   diversifies them in other separate files (eg. tegra18x-mce and
   tegra19x-mce); Shares its header file with other user drivers.
2) Adds a new tegra18x-mce by moving T18x specific functions out
   of the core tegra-mce driver. Renames t19x_mce to tegra19x-mce.
   Allows both of SoC specific drivers and their header files to
   be shared and accessed by tegra-mce core driver only.
3) Adds tegra_ prefixed cache functions by calling different cache
   functions based on SoC ID, instead of overriding default ones.
4) Falls back to default arm64 VA cache functions for T19x instead
   of calling T18x roc functions, since it does not make sense to
   call T18x cache functions on a T19x platform.
5) Raises WARN_ON() if cache flush/clean fails.
6) Fix retry issue in t19x cache functions as retry would be -1
   after the original WARN_ONCE(retry-- == 0).

Bug 200363383
Bug 2017226

Change-Id: I67344ee48cc4faebef91fb7f8cf3b0bd3d4a7cec
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1603072
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>