platform: tegra: remove tegra210-mc-clk-pd domain
Remove tegra210-mc-clk-pd power domain. The domain was previously used
to indicate if MC clock can be turned off. However, there has not been
an implementation to do that in a while so the domain has been
NOP. Hence, remove the code for simplicity and cleaniness.

Bug 1853511

Change-Id: I1f89a1ae685e4b2790cba5e79d61af72bd3a638e
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1605180
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>