dt: remove tegra210-mc-clk-pd power domain
Remove tegra210-mc-clk-pd power domain. The domain was previously used
to indicate if MC clock can be turned off. However, there has not been
an implementation to do that in a while so the domain has been
NOP. Hence, remove the code for simplicity and cleaniness.

Bug 1853511

Change-Id: Idbcab97e82c3bac919c8dd1b76d3d518411dc969
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1605181
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>