ethernet: eqos: check TD bit for normal descriptor
Driver reads RS1V and TSA bits from normal
descriptor to decide on the availability of the
timestamp. After that driver will go and check
for OWN and CTXT bits from context descriptor to
get the valid timestamp.

Sometimes timestamp in context descriptor is
getting dropped due to fifo overflow in other
queues. In this case also driver go and check
for OWN and CTXT bits in context descriptor.
Actually at this point next descriptor is not
a context descriptor but the driver go and read
for those bits.

It will be in loop for checking OWN and CTXT bits
for that context descriptor thinking that DMA
was busy for setting/updating conext descriptor.

Fix it by reading Timestamp dropped bit as well
to decide on the availability of the timestamp.

If the TD bit set in normal descriptor then next
descriptor is not context descriptor.

Bug 200367431

Change-Id: I5262f0b081283fdc1c81f93100fc683e1fee5b9c
Signed-off-by: Bhadram Varka <vbhadram@nvidia.com>
(cherry picked from commit 4e30f7d59ac7d9a40ad5d8e26ff66a573cb038aa)
Reviewed-on: https://git-master.nvidia.com/r/1614045
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>