soc/tgera: pmc: Convert dev_info to dev_err for DPD register dump
GVS system reporting the error when enabling/disabling the
DPD registers. For further debugging, it is needed to dump
DPD REQ and status registers. By dumping via dev_info is
not getting logged in GVS console log and hence changing
it as the dev_err.

Bug 2007223

Change-Id: I1740f893304f2a36e9f7f267c778140259f35897
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1610254
Reviewed-by: Pavan Kunapuli <pkunapuli@nvidia.com>
GVS: Gerrit_Virtual_Submit