t19x: soc: add initial SOC display IMP settings
This patch adds the initial SOC-level display IMP
settings to DT, which will be programmed at boot.
The SOC-level settings are not tuned for any specific
usecases or platforms, and are only there so that
display is functional. These settings can be
overriden by downstream platforms if more
appropriate ones are required.

TDS-3046

Change-Id: Iadcd9465284e7296b5c0ff83c01fe4ce5b51107a
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1610573
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Vishnu Reddy Mandalapu <vmandalapu@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>