mttcan: Remove ptp notifier
None of the driver is using ptp notifier infrastructure. In its current
form, it creates a problem when PTP is not running, then we are not
unregistering the PTP notifier. This leads to kernel Hang. So removing
the unnecessary use of PTP notifier from mttcan driver.

Bug 2022870

Change-Id: Ie5e2a8e5f857e09b8e3d103badcf2bec0eabc26d
Signed-off-by: Abhijit <abhijit@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1611288
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>