ARM64: t210: jetson-cv: Set CYA[0] for I2S4/5
When the I2S5 interface is configured as a bit-clock and frame-sync
slave on Jetson TX1, when configuring the I2S interface, the software
reset for I2S5 fails with the following error ...

 tegra210-i2s tegra210-i2s.4: Failed at I2S4_RX sw reset
 tegra210-i2s tegra210-i2s.4: ASoC: PRE_PMU: I2S5 DAP RX event failed: -22

On Jetson TX1, the I2S5 pins are available only on the I2S5B pads and
per the comment from the following bug, when using I2S5 on these pads
it is necessary to set bit-0 in the I2S5 CYA register. This fixes the
above problem and so ensure that this bit is set for I2S5.

http://nvbugs/1657504/7

Please note that although it was not necessary to set bit-0 in the I2S5
CYA register for I2S5 to function as bit-clock and frame-sync master
when using the I2S5B pads, I have verified that with having this bit
set, I2S5 can still operate without any issues as a master on the I2S5B
pads.

Furthermore, the I2S4 pins are also only available on the I2S4B pads on
Jetson TX1 and so it is also necessary to set bit-0 in the I2S4 CYA
register for this interface as well.

Bug 2029512

Change-Id: I59eb2ea619c3bb70eda4e5672d2b551940174117
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1616462
(cherry picked from commit 70a3f6df02290abe83d5514fd05aad8c5f677399)
Reviewed-on: https://git-master.nvidia.com/r/1642319
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>