dts: soc: t19x: Add interrupt parent for rtc
This patch adds the interrupt parent for RTC node as pm_irq. This
addition helps in treating RTC is a wake event. When system suspend is
triggered, we ensure that the wake mask is programmed to treat RTC as a
tier 2 event and then enable the wake event.

Bug 1964558

Change-Id: Iaf5ee6ff00a313bccad799a32102706830073106
Signed-off-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1614511
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Stefan Kristiansson <stefank@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>