firmware: tegra: fix channel allocation leak
BPMP-FW mail handling keeps track of available threaded channels with
semaphore. In case there is an error in the lower layer of
transmission, the mail handling code errors out early and ends up
"leaking" a semaphore. This can be seen as continuous spew in the
kernel log when bpmp is attempted to access:

[  423.228373] bpmp_write_threaded_ch() down_timeout return -62
[  423.234063] tch_free 0xf to_complete 0x0
[  424.252346] bpmp_write_threaded_ch() down_timeout return -62
[  424.258007] tch_free 0xf to_complete 0x0
[  425.276345] bpmp_write_threaded_ch() down_timeout return -62
[  425.282007] tch_free 0xf to_complete 0x0

Fix the situation by moving the semaphore handling to same function
and ensure acquiring and releasing is always balanced.

Bug 200371530

Change-Id: Ifc5e2396af0cafbd4e5990909b066a5429b1c436
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1647955
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
Tested-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>