l4t: video: tegra: dc: update dc mode selection
In case of L4T, dc mode is set during HDMI init. Mode specified
in DT was not considered as part of this mode selection.
Also, the mode selection is not in accordance with standard HDMI
mode selection policy and there is a lot of duplication of code.
Hence, update the mode selection logic to use DT mode and
refactor the code to adhere to standard policy.

bug 200354689

Change-Id: Iaf82400ecdf1809de17cb7cae73cdbea372a9cfb
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1614115
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>