video: tegra: host: pva: Enable dynamic PM
This change enables dynamic power management for PVAs on silicon.
If there are no users for a PVA for 500ms, the PVA clocks are
disabled and the power domain is powered down. If there are no
longer users for CV rail (e.g. DLA, CVNAS), also the rail gets
powered down.

JIRA PVA-102

Change-Id: If74f0d57b89a61eb3d1e16e6fbc19a6ff8a68b83
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1642225
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>