dc: crc: Store read CRCs if status is valid
When reading CRCs (COMP, RG), mark them as valid if the "valid" bit
of the CRCA register is set, regardless of the error status encoded
in the rest of the CRCA register. The old behavior was to invalidate
the CRCs when there was an error, regardless of the "valid" bit
state.

Bug 2030290

Change-Id: Ia6ec5dcca0dc4c0e3184c6a5a5b004c2470af290
Signed-off-by: Ivan Raul Guadarrama <iguadarrama@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1618152
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Tested-by: Antoine Chauveau <achauveau@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>