clk: tegra210: Do not init clock rate of pll_d2
The pll_d2 clock is used for tegradc -- display on T210. This
clock is usually configured in the bootloader to support early
display.

When kernel takes over the framebuffer from the bootloader, it
should keep the same clock rate for pll_d2 in order to remain
the same pixel clock.

So this patch removes the pll_d2 from the t210_init_table. The
display driver should take the responsibility to configure the
rate according to the display resolution for initialization.

Bug 200369159

Change-Id: I4a3107a3a32e3a729701f8d69fa218bef37685a3
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1618902
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Tested-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: Daniel Fu <danifu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>