nvlink: replace mc reg read/write APIs
Currently the MC driver is disabled for T194. Therefore we can't use the
MC driver's register read/write APIs. The WAR for this is to introduce a
temporary API to access MC registers. This WAR will be removed once the
MC driver gets enabled for T194.

Change-Id: Ia6d6501532e9def6865a8fcfd4960bc4e255d80a
Signed-off-by: Adeel Raza <araza@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1621256
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vijayakumar Subbu <vsubbu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>