t194: cache: Get instead of Set in Hypervisor mode
In Virtualization scenario, RM server programs
the cache way setting and other VMs are not allowed
to write to CACHE_CONTROL registers.

In Linux, userspace might need to know the cache way
allocation and hence the driver is modified to provide
the same by reading the CACHE_CONTROL registers when
hypervisor mode is true.

Jira SSV-1116

Change-Id: If02bf01854adc81345298e5dca48b98c907c24e2
Signed-off-by: Ashok Kumar Sekar <akumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1642726
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>