ntb: Enable switchtec-kernel driver for RT kernel
This change is a manual rebase/merge of patch:
https://www.spinics.net/lists/linux-pci/msg66170.html
which resolves compilation error of switchtec driver
with RT kernel.

Details:
The poll callback is using completion's wait_queue_head_t
member and puts it in poll_wait() so the poll() caller gets
a wakeup after command completed. This does not work on RT
because we don't have a wait_queue_head_t in driver completion
implementation.

Instead of using the completion use waitqueue with a status flag.

JIRA ESDP-1089
JIRA EsDP-1090

Change-Id: If26ead750e75ab4162c7c417349b71cd2e34413e
Signed-off-by: Abhinav Site <asite@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1624712
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>